requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2024.2
